requirements.txt
whoop-data
python-dotenv